+++ ADMF/functions/Test-AdmfDC.ps1
@@ -46,7 +46,7 @@
 	{
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 		$originalArgument = Invoke-PreCredentialProvider @parameters -ProviderName $CredentialProvider -Parameter $parameters -Cmdlet $PSCmdlet
-		try { $parameters.Server = Resolve-DomainController @parameters -ErrorAction Stop }
+		try { $parameters.Server = Resolve-DomainController @parameters -ErrorAction Stop -Confirm:$false }
 		catch
 		{
 			Invoke-PostCredentialProvider -ProviderName $CredentialProvider -Server $originalArgument.Server -Credential $originalArgument.Credential -Cmdlet $PSCmdlet